[REFACTOR/#391] 탐색뷰 분리 headerSection 및 FilterSection
